LayoutTests/fast/viewport/viewport-45.html

<head>
    <title>Viewport META with "minimum-scale=0, maximum-scale=0".</title>
    <meta name="viewport" content="minimum-scale=0, maximum-scale=0">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
