Safari-611.1.10/LayoutTests/fast/viewport/viewport-45-expected.txt

ALERT: viewport size 3200x3520 scale 0.1 with limits [0.1, 0.1] and userScalable true
